Move event file name and data dir into utils header
